Added logging support and changed `Concept.exactMatches` datatype
`Concept.exactMatches` is now a composite of `collections.Mapping` and
`collections.MutableSet`. The mapping operations provide access and
the set operations provide mutability. A mutable dictionary is not
considered appropriate as when setting values the key could differ
from the value of the `Concept.uri` (from which the key should be
derived).